created a new file, prompt.js. updated code and cleaned up.
