+++ public/js/startgame.js
@@ -1,0 +1,7 @@
+const startButton = document.getElementById('startbtn');
+
+startButton.addEventListener('click', startGame);
+
+function startGame() {
+    location.assign('/prompt.html');
+}
