+++ public/prompt.html
@@ -26,7 +26,7 @@ background-size: 100% 150%;">
         <h2 id='time'></h2>
         <p><a href="/prompt">begin</a></p>
     </div>
-    <script src="js/quest.js"></script>
+    <script src="/js/prompt.js"></script>
 </body>
 
 </html>
