+++ public/startgame.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <link href="stylesheets/startgame.css" rel="stylesheet">
-  <script src="game.js">S</script>
+
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
   <title>Adventure game</title>
 </head>
@@ -28,7 +28,7 @@
       <button id="startbtn" class="btn btn-dark btn-lg">Start</button>
     </div>
   </div>
-  <script src="js/quest.js"></script>
+  <script src="/js/startgame.js"></script>
 </body>
 
 </html>
